break out AllocationRequestSerializer
