+++ api/serializers.py
@@ -1,25 +1,12 @@
 from core.models.quota import Quota
-from core.models.allocation import Allocation
 from core.models.identity import Identity
-from core.models.request import AllocationRequest, QuotaRequest, StatusType
+from core.models.request import QuotaRequest, StatusType
 from core.models.user import AtmosphereUser
 
 from rest_framework import serializers
 
 
 # Serializers
-class AllocationRequestSerializer(serializers.ModelSerializer):
-    id = serializers.CharField(read_only=True, source="uuid")
-    created_by = serializers.SlugRelatedField(
-        slug_field='username', source='created_by', read_only=True)
-    status = serializers.SlugRelatedField(
-        slug_field='name', source='status', read_only=True)
-
-    class Meta:
-        model = AllocationRequest
-        exclude = ('uuid', 'membership')
-
-
 class QuotaSerializer(serializers.ModelSerializer):
     class Meta:
         model = Quota
