+++ api/serializers/__init__.py
@@ -40,3 +40,4 @@ from .provider_type_serializer import ProviderTypeSerializer
 from .tag_serializer import TagSerializer
 from .instance_status_history_serializer import InstanceStatusHistorySerializer
 from .allocation_serializer import AllocationSerializer
+from .allocation_request_serializer import AllocationRequestSerializer
