+++ api/serializers/allocation_request_serializer.py
@@ -1,0 +1,14 @@
+from core.models.request import AllocationRequest
+from rest_framework import serializers
+
+
+class AllocationRequestSerializer(serializers.ModelSerializer):
+    id = serializers.CharField(read_only=True, source="uuid")
+    created_by = serializers.SlugRelatedField(
+        slug_field='username', source='created_by', read_only=True)
+    status = serializers.SlugRelatedField(
+        slug_field='name', source='status', read_only=True)
+
+    class Meta:
+        model = AllocationRequest
+        exclude = ('uuid', 'membership')
